abstract `MLP` into a separate `mlp` module for use in other areas of `biogtr`
