+++ biogtr/models/__init__.py
@@ -1,6 +1,7 @@
 """Model architectures and layers."""
 
-from .attention_head import MLP, ATTWeightHead
+from .attention_head import ATTWeightHead
+
 from .embedding import Embedding
 from .transformer import Transformer
 from .visual_encoder import VisualEncoder
